+++ tests/Driver/WslNotifySendDriverTest.php
@@ -13,6 +13,7 @@ namespace Joli\JoliNotif\tests\Driver;
 
 use Joli\JoliNotif\Driver\DriverInterface;
 use Joli\JoliNotif\Driver\WslNotifySendDriver;
+use Psr\Log\NullLogger;
 
 class WslNotifySendDriverTest extends AbstractDriverTestCase
 {
@@ -37,7 +38,7 @@ class WslNotifySendDriverTest extends AbstractDriverTestCase
 
     protected function getDriver(): WslNotifySendDriver
     {
-        return new WslNotifySendDriver();
+        return new WslNotifySendDriver(new NullLogger());
     }
 
     protected function getExpectedCommandLineForNotification(): string